Factorize job events.
