+++ src/Illuminate/Queue/Events/JobProcessing.php
@@ -2,32 +2,7 @@
 
 namespace Illuminate\Queue\Events;
 
-class JobProcessing
+class JobProcessing extends AbstractJobEvent
 {
-    /**
-     * The connection name.
-     *
-     * @var string
-     */
-    public $connectionName;
-
-    /**
-     * The job instance.
-     *
-     * @var \Illuminate\Contracts\Queue\Job
-     */
-    public $job;
-
-    /**
-     * Create a new event instance.
-     *
-     * @param  string  $connectionName
-     * @param  \Illuminate\Contracts\Queue\Job  $job
-     * @return void
-     */
-    public function __construct($connectionName, $job)
-    {
-        $this->job = $job;
-        $this->connectionName = $connectionName;
-    }
+    //
 }
